Module Sarek_vulkan.Vulkan_types

Basic Types

type vk_handle = Unsigned.uint64

Vulkan handles are 64-bit on 64-bit platforms

val vk_handle : vk_handle Ctypes.typ

type vk_non_dispatchable_handle = Unsigned.uint64

Non-dispatchable handles (pointers on 64-bit, uint64 on 32-bit)

val vk_non_dispatchable_handle : vk_non_dispatchable_handle Ctypes.typ

type vk_device_size = Unsigned.uint64

VkDeviceSize for memory sizes

val vk_device_size : vk_device_size Ctypes.typ

type vk_flags = Unsigned.uint32

VkFlags for bitmasks

val vk_flags : vk_flags Ctypes.typ

type vk_bool32 = Unsigned.uint32

VkBool32

val vk_bool32 : vk_bool32 Ctypes.typ

val vk_true : Unsigned.UInt32.t

val vk_false : Unsigned.UInt32.t

Opaque Handle Types

type vk_instance

Dispatchable handles (pointer-sized)

val vk_instance : vk_instance Ctypes.structure Ctypes.typ

val vk_instance_ptr : vk_instance Ctypes.structure Ctypes_static.ptr Ctypes.typ

type vk_physical_device

val vk_physical_device : vk_physical_device Ctypes.structure Ctypes.typ

val vk_physical_device_ptr : vk_physical_device Ctypes.structure Ctypes_static.ptr Ctypes.typ

type vk_device

val vk_device : vk_device Ctypes.structure Ctypes.typ

val vk_device_ptr : vk_device Ctypes.structure Ctypes_static.ptr Ctypes.typ

type vk_queue

val vk_queue : vk_queue Ctypes.structure Ctypes.typ

val vk_queue_ptr : vk_queue Ctypes.structure Ctypes_static.ptr Ctypes.typ

type vk_command_buffer

val vk_command_buffer : vk_command_buffer Ctypes.structure Ctypes.typ

val vk_command_buffer_ptr : vk_command_buffer Ctypes.structure Ctypes_static.ptr Ctypes.typ

type vk_buffer = vk_non_dispatchable_handle

Non-dispatchable handles

val vk_buffer : vk_buffer Ctypes.typ

type vk_device_memory = vk_non_dispatchable_handle

val vk_device_memory : vk_device_memory Ctypes.typ

type vk_shader_module = vk_non_dispatchable_handle

val vk_shader_module : vk_shader_module Ctypes.typ

type vk_pipeline = vk_non_dispatchable_handle

val vk_pipeline : vk_pipeline Ctypes.typ

type vk_pipeline_layout = vk_non_dispatchable_handle

val vk_pipeline_layout : vk_pipeline_layout Ctypes.typ

type vk_descriptor_set_layout = vk_non_dispatchable_handle

val vk_descriptor_set_layout : vk_descriptor_set_layout Ctypes.typ

type vk_descriptor_pool = vk_non_dispatchable_handle

val vk_descriptor_pool : vk_descriptor_pool Ctypes.typ

type vk_descriptor_set = vk_non_dispatchable_handle

val vk_descriptor_set : vk_descriptor_set Ctypes.typ

type vk_command_pool = vk_non_dispatchable_handle

val vk_command_pool : vk_command_pool Ctypes.typ

type vk_fence = vk_non_dispatchable_handle

val vk_fence : vk_fence Ctypes.typ

type vk_semaphore = vk_non_dispatchable_handle

val vk_semaphore : vk_semaphore Ctypes.typ

type vk_event = vk_non_dispatchable_handle

val vk_event : vk_event Ctypes.typ

type vk_pipeline_cache = vk_non_dispatchable_handle

val vk_pipeline_cache : vk_pipeline_cache Ctypes.typ

Result Codes

type vk_result =

| VK_SUCCESS

| VK_NOT_READY

| VK_TIMEOUT

| VK_EVENT_SET

| VK_EVENT_RESET

| VK_INCOMPLETE

| VK_ERROR_OUT_OF_HOST_MEMORY

| VK_ERROR_OUT_OF_DEVICE_MEMORY

| VK_ERROR_INITIALIZATION_FAILED

| VK_ERROR_DEVICE_LOST

| VK_ERROR_MEMORY_MAP_FAILED

| VK_ERROR_LAYER_NOT_PRESENT

| VK_ERROR_EXTENSION_NOT_PRESENT

| VK_ERROR_FEATURE_NOT_PRESENT

| VK_ERROR_INCOMPATIBLE_DRIVER

| VK_ERROR_TOO_MANY_OBJECTS

| VK_ERROR_FORMAT_NOT_SUPPORTED

| VK_ERROR_FRAGMENTED_POOL

| VK_ERROR_UNKNOWN of int

val vk_result_of_int : int -> vk_result

val int_of_vk_result : vk_result -> int

val vk_result : vk_result Ctypes.typ

val string_of_vk_result : vk_result -> string

Structure Types

type vk_structure_type = int

val vk_structure_type_application_info : int

val vk_structure_type_instance_create_info : int

val vk_structure_type_device_queue_create_info : int

val vk_structure_type_device_create_info : int

val vk_structure_type_submit_info : int

val vk_structure_type_memory_allocate_info : int

val vk_structure_type_buffer_create_info : int

val vk_structure_type_shader_module_create_info : int

val vk_structure_type_compute_pipeline_create_info : int

val vk_structure_type_pipeline_layout_create_info : int

val vk_structure_type_descriptor_set_layout_create_info : int

val vk_structure_type_descriptor_pool_create_info : int

val vk_structure_type_descriptor_set_allocate_info : int

val vk_structure_type_write_descriptor_set : int

val vk_structure_type_fence_create_info : int

val vk_structure_type_command_pool_create_info : int

val vk_structure_type_command_buffer_allocate_info : int

val vk_structure_type_command_buffer_begin_info : int

val vk_structure_type_pipeline_shader_stage_create_info : int

Queue Family Properties

val vk_queue_compute_bit : int

val vk_queue_transfer_bit : int

Memory Property Flags

val vk_memory_property_device_local_bit : int

val vk_memory_property_host_visible_bit : int

val vk_memory_property_host_coherent_bit : int

val vk_memory_property_host_cached_bit : int

Buffer Usage Flags

val vk_buffer_usage_transfer_src_bit : int

val vk_buffer_usage_transfer_dst_bit : int

val vk_buffer_usage_storage_buffer_bit : int

Descriptor Types

val vk_descriptor_type_storage_buffer : int

Shader Stage Flags

val vk_shader_stage_compute_bit : int

Pipeline Stage Flags

val vk_pipeline_stage_host_bit : int

val vk_pipeline_stage_compute_shader_bit : int

Access Flags

val vk_access_host_write_bit : int

val vk_access_host_read_bit : int

val vk_access_shader_read_bit : int

val vk_access_shader_write_bit : int

Misc Constants

val vk_queue_family_ignored : int

val vk_structure_type_buffer_memory_barrier : int

Pipeline Bind Points

val vk_pipeline_bind_point_compute : int

Command Buffer Level

val vk_command_buffer_level_primary : int

Command Buffer Usage Flags

val vk_command_buffer_usage_one_time_submit_bit : int

Fence Create Flags

val vk_fence_create_signaled_bit : int

Structures

type vk_application_info

val vk_application_info : vk_application_info Ctypes.structure Ctypes.typ

val app_info_sType : (Unsigned.uint32, vk_application_info Ctypes.structure) Ctypes.field

val app_info_pNext : (unit Ctypes_static.ptr, vk_application_info Ctypes.structure) Ctypes.field

val app_info_pApplicationName : (char Ctypes_static.ptr, vk_application_info Ctypes.structure) Ctypes.field

val app_info_applicationVersion : (Unsigned.uint32, vk_application_info Ctypes.structure) Ctypes.field

val app_info_pEngineName : (char Ctypes_static.ptr, vk_application_info Ctypes.structure) Ctypes.field

val app_info_engineVersion : (Unsigned.uint32, vk_application_info Ctypes.structure) Ctypes.field

val app_info_apiVersion : (Unsigned.uint32, vk_application_info Ctypes.structure) Ctypes.field

type vk_instance_create_info

val vk_instance_create_info : vk_instance_create_info Ctypes.structure Ctypes.typ

val inst_create_sType : (Unsigned.uint32, vk_instance_create_info Ctypes.structure) Ctypes.field

val inst_create_pNext : (unit Ctypes_static.ptr, vk_instance_create_info Ctypes.structure) Ctypes.field

val inst_create_flags : (vk_flags, vk_instance_create_info Ctypes.structure) Ctypes.field

val inst_create_pApplicationInfo : (vk_application_info Ctypes.structure Ctypes_static.ptr, vk_instance_create_info Ctypes.structure) Ctypes.field

val inst_create_enabledLayerCount : (Unsigned.uint32, vk_instance_create_info Ctypes.structure) Ctypes.field

val inst_create_ppEnabledLayerNames : (string Ctypes_static.ptr, vk_instance_create_info Ctypes.structure) Ctypes.field

val inst_create_enabledExtensionCount : (Unsigned.uint32, vk_instance_create_info Ctypes.structure) Ctypes.field

val inst_create_ppEnabledExtensionNames : (string Ctypes_static.ptr, vk_instance_create_info Ctypes.structure) Ctypes.field

type vk_physical_device_properties

val vk_physical_device_properties : vk_physical_device_properties Ctypes.structure Ctypes.typ

val phys_props_apiVersion : (Unsigned.uint32, vk_physical_device_properties Ctypes.structure) Ctypes.field

val phys_props_driverVersion : (Unsigned.uint32, vk_physical_device_properties Ctypes.structure) Ctypes.field

val phys_props_vendorID : (Unsigned.uint32, vk_physical_device_properties Ctypes.structure) Ctypes.field

val phys_props_deviceID : (Unsigned.uint32, vk_physical_device_properties Ctypes.structure) Ctypes.field

val phys_props_deviceType : (Unsigned.uint32, vk_physical_device_properties Ctypes.structure) Ctypes.field

val phys_props_deviceName : (char Ctypes_static.carray, vk_physical_device_properties Ctypes.structure) Ctypes.field

val phys_props_padding : (char Ctypes_static.carray, vk_physical_device_properties Ctypes.structure) Ctypes.field

type vk_physical_device_features

val vk_physical_device_features_field_count : int

val shader_float64_field_index : int

val shader_int64_field_index : int

val vk_physical_device_features : vk_physical_device_features Ctypes.structure Ctypes.typ

val phys_features_bools : (Unsigned.uint32 Ctypes_static.carray, vk_physical_device_features Ctypes.structure) Ctypes.field

type vk_queue_family_properties

val vk_queue_family_properties : vk_queue_family_properties Ctypes.structure Ctypes.typ

val queue_family_queueFlags : (vk_flags, vk_queue_family_properties Ctypes.structure) Ctypes.field

val queue_family_queueCount : (Unsigned.uint32, vk_queue_family_properties Ctypes.structure) Ctypes.field

val queue_family_timestampValidBits : (Unsigned.uint32, vk_queue_family_properties Ctypes.structure) Ctypes.field

val queue_family_minImageTransferGranularity_width : (Unsigned.uint32, vk_queue_family_properties Ctypes.structure) Ctypes.field

val queue_family_minImageTransferGranularity_height : (Unsigned.uint32, vk_queue_family_properties Ctypes.structure) Ctypes.field

val queue_family_minImageTransferGranularity_depth : (Unsigned.uint32, vk_queue_family_properties Ctypes.structure) Ctypes.field

type vk_device_queue_create_info

val vk_device_queue_create_info : vk_device_queue_create_info Ctypes.structure Ctypes.typ

val dev_queue_create_sType : (Unsigned.uint32, vk_device_queue_create_info Ctypes.structure) Ctypes.field

val dev_queue_create_pNext : (unit Ctypes_static.ptr, vk_device_queue_create_info Ctypes.structure) Ctypes.field

val dev_queue_create_flags : (vk_flags, vk_device_queue_create_info Ctypes.structure) Ctypes.field

val dev_queue_create_queueFamilyIndex : (Unsigned.uint32, vk_device_queue_create_info Ctypes.structure) Ctypes.field

val dev_queue_create_queueCount : (Unsigned.uint32, vk_device_queue_create_info Ctypes.structure) Ctypes.field

val dev_queue_create_pQueuePriorities : (float Ctypes_static.ptr, vk_device_queue_create_info Ctypes.structure) Ctypes.field

type vk_device_create_info

val vk_device_create_info : vk_device_create_info Ctypes.structure Ctypes.typ

val dev_create_sType : (Unsigned.uint32, vk_device_create_info Ctypes.structure) Ctypes.field

val dev_create_pNext : (unit Ctypes_static.ptr, vk_device_create_info Ctypes.structure) Ctypes.field

val dev_create_flags : (vk_flags, vk_device_create_info Ctypes.structure) Ctypes.field

val dev_create_queueCreateInfoCount : (Unsigned.uint32, vk_device_create_info Ctypes.structure) Ctypes.field

val dev_create_pQueueCreateInfos : (vk_device_queue_create_info Ctypes.structure Ctypes_static.ptr, vk_device_create_info Ctypes.structure) Ctypes.field

val dev_create_enabledLayerCount : (Unsigned.uint32, vk_device_create_info Ctypes.structure) Ctypes.field

val dev_create_ppEnabledLayerNames : (string Ctypes_static.ptr, vk_device_create_info Ctypes.structure) Ctypes.field

val dev_create_enabledExtensionCount : (Unsigned.uint32, vk_device_create_info Ctypes.structure) Ctypes.field

val dev_create_ppEnabledExtensionNames : (string Ctypes_static.ptr, vk_device_create_info Ctypes.structure) Ctypes.field

val dev_create_pEnabledFeatures : (unit Ctypes_static.ptr, vk_device_create_info Ctypes.structure) Ctypes.field

type vk_memory_allocate_info

val vk_memory_allocate_info : vk_memory_allocate_info Ctypes.structure Ctypes.typ

val mem_alloc_sType : (Unsigned.uint32, vk_memory_allocate_info Ctypes.structure) Ctypes.field

val mem_alloc_pNext : (unit Ctypes_static.ptr, vk_memory_allocate_info Ctypes.structure) Ctypes.field

val mem_alloc_allocationSize : (vk_device_size, vk_memory_allocate_info Ctypes.structure) Ctypes.field

val mem_alloc_memoryTypeIndex : (Unsigned.uint32, vk_memory_allocate_info Ctypes.structure) Ctypes.field

type vk_buffer_create_info

val vk_buffer_create_info : vk_buffer_create_info Ctypes.structure Ctypes.typ

val buf_create_sType : (Unsigned.uint32, vk_buffer_create_info Ctypes.structure) Ctypes.field

val buf_create_pNext : (unit Ctypes_static.ptr, vk_buffer_create_info Ctypes.structure) Ctypes.field

val buf_create_flags : (vk_flags, vk_buffer_create_info Ctypes.structure) Ctypes.field

val buf_create_size : (vk_device_size, vk_buffer_create_info Ctypes.structure) Ctypes.field

val buf_create_usage : (vk_flags, vk_buffer_create_info Ctypes.structure) Ctypes.field

val buf_create_sharingMode : (Unsigned.uint32, vk_buffer_create_info Ctypes.structure) Ctypes.field

val buf_create_queueFamilyIndexCount : (Unsigned.uint32, vk_buffer_create_info Ctypes.structure) Ctypes.field

val buf_create_pQueueFamilyIndices : (Unsigned.uint32 Ctypes_static.ptr, vk_buffer_create_info Ctypes.structure) Ctypes.field

type vk_memory_requirements

val vk_memory_requirements : vk_memory_requirements Ctypes.structure Ctypes.typ

val mem_req_size : (vk_device_size, vk_memory_requirements Ctypes.structure) Ctypes.field

val mem_req_alignment : (vk_device_size, vk_memory_requirements Ctypes.structure) Ctypes.field

val mem_req_memoryTypeBits : (Unsigned.uint32, vk_memory_requirements Ctypes.structure) Ctypes.field

type vk_memory_type

val vk_memory_type : vk_memory_type Ctypes.structure Ctypes.typ

val mem_type_propertyFlags : (vk_flags, vk_memory_type Ctypes.structure) Ctypes.field

val mem_type_heapIndex : (Unsigned.uint32, vk_memory_type Ctypes.structure) Ctypes.field

type vk_memory_heap

val vk_memory_heap : vk_memory_heap Ctypes.structure Ctypes.typ

val mem_heap_size : (vk_device_size, vk_memory_heap Ctypes.structure) Ctypes.field

val mem_heap_flags : (vk_flags, vk_memory_heap Ctypes.structure) Ctypes.field

type vk_physical_device_memory_properties

val vk_physical_device_memory_properties : vk_physical_device_memory_properties Ctypes.structure Ctypes.typ

val mem_props_memoryTypeCount : (Unsigned.uint32, vk_physical_device_memory_properties Ctypes.structure) Ctypes.field

val mem_props_memoryTypes : (vk_memory_type Ctypes.structure Ctypes_static.carray, vk_physical_device_memory_properties Ctypes.structure) Ctypes.field

val mem_props_memoryHeapCount : (Unsigned.uint32, vk_physical_device_memory_properties Ctypes.structure) Ctypes.field

val mem_props_memoryHeaps : (vk_memory_heap Ctypes.structure Ctypes_static.carray, vk_physical_device_memory_properties Ctypes.structure) Ctypes.field

type vk_shader_module_create_info

val vk_shader_module_create_info : vk_shader_module_create_info Ctypes.structure Ctypes.typ

val shader_create_sType : (Unsigned.uint32, vk_shader_module_create_info Ctypes.structure) Ctypes.field

val shader_create_pNext : (unit Ctypes_static.ptr, vk_shader_module_create_info Ctypes.structure) Ctypes.field

val shader_create_flags : (vk_flags, vk_shader_module_create_info Ctypes.structure) Ctypes.field

val shader_create_codeSize : (Unsigned.size_t, vk_shader_module_create_info Ctypes.structure) Ctypes.field

val shader_create_pCode : (Unsigned.uint32 Ctypes_static.ptr, vk_shader_module_create_info Ctypes.structure) Ctypes.field

type vk_command_pool_create_info

val vk_command_pool_create_info : vk_command_pool_create_info Ctypes.structure Ctypes.typ

val cmd_pool_create_sType : (Unsigned.uint32, vk_command_pool_create_info Ctypes.structure) Ctypes.field

val cmd_pool_create_pNext : (unit Ctypes_static.ptr, vk_command_pool_create_info Ctypes.structure) Ctypes.field

val cmd_pool_create_flags : (vk_flags, vk_command_pool_create_info Ctypes.structure) Ctypes.field

val cmd_pool_create_queueFamilyIndex : (Unsigned.uint32, vk_command_pool_create_info Ctypes.structure) Ctypes.field

type vk_command_buffer_allocate_info

val vk_command_buffer_allocate_info : vk_command_buffer_allocate_info Ctypes.structure Ctypes.typ

val cmd_buf_alloc_sType : (Unsigned.uint32, vk_command_buffer_allocate_info Ctypes.structure) Ctypes.field

val cmd_buf_alloc_pNext : (unit Ctypes_static.ptr, vk_command_buffer_allocate_info Ctypes.structure) Ctypes.field

val cmd_buf_alloc_commandPool : (vk_command_pool, vk_command_buffer_allocate_info Ctypes.structure) Ctypes.field

val cmd_buf_alloc_level : (Unsigned.uint32, vk_command_buffer_allocate_info Ctypes.structure) Ctypes.field

val cmd_buf_alloc_commandBufferCount : (Unsigned.uint32, vk_command_buffer_allocate_info Ctypes.structure) Ctypes.field

type vk_fence_create_info

val vk_fence_create_info : vk_fence_create_info Ctypes.structure Ctypes.typ

val fence_create_sType : (Unsigned.uint32, vk_fence_create_info Ctypes.structure) Ctypes.field

val fence_create_pNext : (unit Ctypes_static.ptr, vk_fence_create_info Ctypes.structure) Ctypes.field

val fence_create_flags : (vk_flags, vk_fence_create_info Ctypes.structure) Ctypes.field

type vk_submit_info

val vk_submit_info : vk_submit_info Ctypes.structure Ctypes.typ

val submit_sType : (Unsigned.uint32, vk_submit_info Ctypes.structure) Ctypes.field

val submit_pNext : (unit Ctypes_static.ptr, vk_submit_info Ctypes.structure) Ctypes.field

val submit_waitSemaphoreCount : (Unsigned.uint32, vk_submit_info Ctypes.structure) Ctypes.field

val submit_pWaitSemaphores : (vk_semaphore Ctypes_static.ptr, vk_submit_info Ctypes.structure) Ctypes.field

val submit_pWaitDstStageMask : (vk_flags Ctypes_static.ptr, vk_submit_info Ctypes.structure) Ctypes.field

val submit_commandBufferCount : (Unsigned.uint32, vk_submit_info Ctypes.structure) Ctypes.field

val submit_pCommandBuffers : (vk_command_buffer Ctypes.structure Ctypes_static.ptr Ctypes_static.ptr, vk_submit_info Ctypes.structure) Ctypes.field

val submit_signalSemaphoreCount : (Unsigned.uint32, vk_submit_info Ctypes.structure) Ctypes.field

val submit_pSignalSemaphores : (vk_semaphore Ctypes_static.ptr, vk_submit_info Ctypes.structure) Ctypes.field

type vk_descriptor_set_layout_binding

VkDescriptorSetLayoutBinding

val vk_descriptor_set_layout_binding : vk_descriptor_set_layout_binding Ctypes.structure Ctypes.typ

val dsl_binding_binding : (Unsigned.uint32, vk_descriptor_set_layout_binding Ctypes.structure) Ctypes.field

val dsl_binding_descriptorType : (Unsigned.uint32, vk_descriptor_set_layout_binding Ctypes.structure) Ctypes.field

val dsl_binding_descriptorCount : (Unsigned.uint32, vk_descriptor_set_layout_binding Ctypes.structure) Ctypes.field

val dsl_binding_stageFlags : (vk_flags, vk_descriptor_set_layout_binding Ctypes.structure) Ctypes.field

val dsl_binding_pImmutableSamplers : (unit Ctypes_static.ptr, vk_descriptor_set_layout_binding Ctypes.structure) Ctypes.field

type vk_descriptor_set_layout_create_info

VkDescriptorSetLayoutCreateInfo

val vk_descriptor_set_layout_create_info : vk_descriptor_set_layout_create_info Ctypes.structure Ctypes.typ

val dsl_create_sType : (Unsigned.uint32, vk_descriptor_set_layout_create_info Ctypes.structure) Ctypes.field

val dsl_create_pNext : (unit Ctypes_static.ptr, vk_descriptor_set_layout_create_info Ctypes.structure) Ctypes.field

val dsl_create_flags : (vk_flags, vk_descriptor_set_layout_create_info Ctypes.structure) Ctypes.field

val dsl_create_bindingCount : (Unsigned.uint32, vk_descriptor_set_layout_create_info Ctypes.structure) Ctypes.field

val dsl_create_pBindings : (vk_descriptor_set_layout_binding Ctypes.structure Ctypes_static.ptr, vk_descriptor_set_layout_create_info Ctypes.structure) Ctypes.field

type vk_push_constant_range

VkPushConstantRange

val vk_push_constant_range : vk_push_constant_range Ctypes.structure Ctypes.typ

val push_const_stageFlags : (Unsigned.uint32, vk_push_constant_range Ctypes.structure) Ctypes.field

val push_const_offset : (Unsigned.uint32, vk_push_constant_range Ctypes.structure) Ctypes.field

val push_const_size : (Unsigned.uint32, vk_push_constant_range Ctypes.structure) Ctypes.field

type vk_buffer_memory_barrier

VkBufferMemoryBarrier

val vk_buffer_memory_barrier : vk_buffer_memory_barrier Ctypes.structure Ctypes.typ

val buf_barrier_sType : (Unsigned.uint32, vk_buffer_memory_barrier Ctypes.structure) Ctypes.field

val buf_barrier_pNext : (unit Ctypes_static.ptr, vk_buffer_memory_barrier Ctypes.structure) Ctypes.field

val buf_barrier_srcAccessMask : (vk_flags, vk_buffer_memory_barrier Ctypes.structure) Ctypes.field

val buf_barrier_dstAccessMask : (vk_flags, vk_buffer_memory_barrier Ctypes.structure) Ctypes.field

val buf_barrier_srcQueueFamilyIndex : (Unsigned.uint32, vk_buffer_memory_barrier Ctypes.structure) Ctypes.field

val buf_barrier_dstQueueFamilyIndex : (Unsigned.uint32, vk_buffer_memory_barrier Ctypes.structure) Ctypes.field

val buf_barrier_buffer : (vk_buffer, vk_buffer_memory_barrier Ctypes.structure) Ctypes.field

val buf_barrier_offset : (vk_device_size, vk_buffer_memory_barrier Ctypes.structure) Ctypes.field

val buf_barrier_size : (vk_device_size, vk_buffer_memory_barrier Ctypes.structure) Ctypes.field

type vk_pipeline_layout_create_info

VkPipelineLayoutCreateInfo

val vk_pipeline_layout_create_info : vk_pipeline_layout_create_info Ctypes.structure Ctypes.typ

val pl_create_sType : (Unsigned.uint32, vk_pipeline_layout_create_info Ctypes.structure) Ctypes.field

val pl_create_pNext : (unit Ctypes_static.ptr, vk_pipeline_layout_create_info Ctypes.structure) Ctypes.field

val pl_create_flags : (vk_flags, vk_pipeline_layout_create_info Ctypes.structure) Ctypes.field

val pl_create_setLayoutCount : (Unsigned.uint32, vk_pipeline_layout_create_info Ctypes.structure) Ctypes.field

val pl_create_pSetLayouts : (vk_descriptor_set_layout Ctypes_static.ptr, vk_pipeline_layout_create_info Ctypes.structure) Ctypes.field

val pl_create_pushConstantRangeCount : (Unsigned.uint32, vk_pipeline_layout_create_info Ctypes.structure) Ctypes.field

val pl_create_pPushConstantRanges : (vk_push_constant_range Ctypes.structure Ctypes_static.ptr, vk_pipeline_layout_create_info Ctypes.structure) Ctypes.field

type vk_pipeline_shader_stage_create_info

VkPipelineShaderStageCreateInfo

val vk_pipeline_shader_stage_create_info : vk_pipeline_shader_stage_create_info Ctypes.structure Ctypes.typ

val shader_stage_sType : (Unsigned.uint32, vk_pipeline_shader_stage_create_info Ctypes.structure) Ctypes.field

val shader_stage_pNext : (unit Ctypes_static.ptr, vk_pipeline_shader_stage_create_info Ctypes.structure) Ctypes.field

val shader_stage_flags : (vk_flags, vk_pipeline_shader_stage_create_info Ctypes.structure) Ctypes.field

val shader_stage_stage : (vk_flags, vk_pipeline_shader_stage_create_info Ctypes.structure) Ctypes.field

val shader_stage_module : (vk_shader_module, vk_pipeline_shader_stage_create_info Ctypes.structure) Ctypes.field

val shader_stage_pName : (char Ctypes_static.ptr, vk_pipeline_shader_stage_create_info Ctypes.structure) Ctypes.field

val shader_stage_pSpecializationInfo : (unit Ctypes_static.ptr, vk_pipeline_shader_stage_create_info Ctypes.structure) Ctypes.field

type vk_compute_pipeline_create_info

VkComputePipelineCreateInfo

val vk_compute_pipeline_create_info : vk_compute_pipeline_create_info Ctypes.structure Ctypes.typ

val compute_pipe_sType : (Unsigned.uint32, vk_compute_pipeline_create_info Ctypes.structure) Ctypes.field

val compute_pipe_pNext : (unit Ctypes_static.ptr, vk_compute_pipeline_create_info Ctypes.structure) Ctypes.field

val compute_pipe_flags : (vk_flags, vk_compute_pipeline_create_info Ctypes.structure) Ctypes.field

val compute_pipe_stage : (vk_pipeline_shader_stage_create_info Ctypes.structure, vk_compute_pipeline_create_info Ctypes.structure) Ctypes.field

val compute_pipe_layout : (vk_pipeline_layout, vk_compute_pipeline_create_info Ctypes.structure) Ctypes.field

val compute_pipe_basePipelineHandle : (vk_pipeline, vk_compute_pipeline_create_info Ctypes.structure) Ctypes.field

val compute_pipe_basePipelineIndex : (int32, vk_compute_pipeline_create_info Ctypes.structure) Ctypes.field

type vk_descriptor_pool_size

VkDescriptorPoolSize

val vk_descriptor_pool_size : vk_descriptor_pool_size Ctypes.structure Ctypes.typ

val pool_size_type : (Unsigned.uint32, vk_descriptor_pool_size Ctypes.structure) Ctypes.field

val pool_size_descriptorCount : (Unsigned.uint32, vk_descriptor_pool_size Ctypes.structure) Ctypes.field

type vk_descriptor_pool_create_info

VkDescriptorPoolCreateInfo

val vk_descriptor_pool_create_info : vk_descriptor_pool_create_info Ctypes.structure Ctypes.typ

val desc_pool_sType : (Unsigned.uint32, vk_descriptor_pool_create_info Ctypes.structure) Ctypes.field

val desc_pool_pNext : (unit Ctypes_static.ptr, vk_descriptor_pool_create_info Ctypes.structure) Ctypes.field

val desc_pool_flags : (vk_flags, vk_descriptor_pool_create_info Ctypes.structure) Ctypes.field

val desc_pool_maxSets : (Unsigned.uint32, vk_descriptor_pool_create_info Ctypes.structure) Ctypes.field

val desc_pool_poolSizeCount : (Unsigned.uint32, vk_descriptor_pool_create_info Ctypes.structure) Ctypes.field

val desc_pool_pPoolSizes : (vk_descriptor_pool_size Ctypes.structure Ctypes_static.ptr, vk_descriptor_pool_create_info Ctypes.structure) Ctypes.field

type vk_descriptor_set_allocate_info

VkDescriptorSetAllocateInfo

val vk_descriptor_set_allocate_info : vk_descriptor_set_allocate_info Ctypes.structure Ctypes.typ

val desc_set_alloc_sType : (Unsigned.uint32, vk_descriptor_set_allocate_info Ctypes.structure) Ctypes.field

val desc_set_alloc_pNext : (unit Ctypes_static.ptr, vk_descriptor_set_allocate_info Ctypes.structure) Ctypes.field

val desc_set_alloc_descriptorPool : (vk_descriptor_pool, vk_descriptor_set_allocate_info Ctypes.structure) Ctypes.field

val desc_set_alloc_descriptorSetCount : (Unsigned.uint32, vk_descriptor_set_allocate_info Ctypes.structure) Ctypes.field

val desc_set_alloc_pSetLayouts : (vk_descriptor_set_layout Ctypes_static.ptr, vk_descriptor_set_allocate_info Ctypes.structure) Ctypes.field

type vk_descriptor_buffer_info

VkDescriptorBufferInfo

val vk_descriptor_buffer_info : vk_descriptor_buffer_info Ctypes.structure Ctypes.typ

val desc_buf_buffer : (vk_buffer, vk_descriptor_buffer_info Ctypes.structure) Ctypes.field

val desc_buf_offset : (vk_device_size, vk_descriptor_buffer_info Ctypes.structure) Ctypes.field

val desc_buf_range : (vk_device_size, vk_descriptor_buffer_info Ctypes.structure) Ctypes.field

type vk_write_descriptor_set

VkWriteDescriptorSet

val vk_write_descriptor_set : vk_write_descriptor_set Ctypes.structure Ctypes.typ

val write_desc_sType : (Unsigned.uint32, vk_write_descriptor_set Ctypes.structure) Ctypes.field

val write_desc_pNext : (unit Ctypes_static.ptr, vk_write_descriptor_set Ctypes.structure) Ctypes.field

val write_desc_dstSet : (vk_descriptor_set, vk_write_descriptor_set Ctypes.structure) Ctypes.field

val write_desc_dstBinding : (Unsigned.uint32, vk_write_descriptor_set Ctypes.structure) Ctypes.field

val write_desc_dstArrayElement : (Unsigned.uint32, vk_write_descriptor_set Ctypes.structure) Ctypes.field

val write_desc_descriptorCount : (Unsigned.uint32, vk_write_descriptor_set Ctypes.structure) Ctypes.field

val write_desc_descriptorType : (Unsigned.uint32, vk_write_descriptor_set Ctypes.structure) Ctypes.field

val write_desc_pImageInfo : (unit Ctypes_static.ptr, vk_write_descriptor_set Ctypes.structure) Ctypes.field

val write_desc_pBufferInfo : (vk_descriptor_buffer_info Ctypes.structure Ctypes_static.ptr, vk_write_descriptor_set Ctypes.structure) Ctypes.field

val write_desc_pTexelBufferView : (unit Ctypes_static.ptr, vk_write_descriptor_set Ctypes.structure) Ctypes.field

type vk_command_buffer_begin_info

VkCommandBufferBeginInfo

val vk_command_buffer_begin_info : vk_command_buffer_begin_info Ctypes.structure Ctypes.typ

val cmd_buf_begin_sType : (Unsigned.uint32, vk_command_buffer_begin_info Ctypes.structure) Ctypes.field

val cmd_buf_begin_pNext : (unit Ctypes_static.ptr, vk_command_buffer_begin_info Ctypes.structure) Ctypes.field

val cmd_buf_begin_flags : (vk_flags, vk_command_buffer_begin_info Ctypes.structure) Ctypes.field

val cmd_buf_begin_pInheritanceInfo : (unit Ctypes_static.ptr, vk_command_buffer_begin_info Ctypes.structure) Ctypes.field

val vk_null_handle : Unsigned.UInt64.t

Null handle constant

type vk_buffer_copy

VkBufferCopy

val vk_buffer_copy : vk_buffer_copy Ctypes.structure Ctypes.typ

val buffer_copy_srcOffset : (vk_device_size, vk_buffer_copy Ctypes.structure) Ctypes.field

val buffer_copy_dstOffset : (vk_device_size, vk_buffer_copy Ctypes.structure) Ctypes.field

val buffer_copy_size : (vk_device_size, vk_buffer_copy Ctypes.structure) Ctypes.field

Extended feature and property queries (backlog-62 slice 2)

Everything below is reached through the pNext chains of VkPhysicalDeviceFeatures2 / VkPhysicalDeviceProperties2, which is the only way to query shaderFloat16, storageBuffer16BitAccess, cooperativeMatrix, the driver identity and the subgroup size. Both entry points are core Vulkan 1.1, and the instance this backend creates requests API 1.2, so neither needs an instance extension.

Two rules the structs below obey, and both were arrived at the hard way:

Every struct is ZEROED and given its sType before the query. A driver that does not recognise an sType leaves the struct untouched, so an uninitialised struct would be read as "feature present" on exactly the devices that lack it.

A struct chained into a query must be at least as large as the driver believes it to be. vk_physical_device_properties above is a TRUNCATED model (it stops after deviceName and pads), so it cannot be embedded in VkPhysicalDeviceProperties2 by value; the properties2 wrapper below carries an opaque blob sized past the real struct instead.

val vk_structure_type_physical_device_features_2 : int

val vk_structure_type_physical_device_properties_2 : int

val vk_structure_type_physical_device_shader_float16_int8_features : int

val vk_structure_type_physical_device_16bit_storage_features : int

backlog-62 slice 3. The three enumerants below are what the INTEGER cooperative-matrix path needs and the f16 path did not.

Measured, not guessed: glslc --target-env=vulkan1.3 on a 16x16x16 u8 x u8 -> s32 coopMatMulAdd emits exactly OpCapability Int8, OpCapability StorageBuffer8BitAccess, OpCapability VulkanMemoryModel and OpCapability CooperativeMatrixKHR. Slice 2 enables only the last. The first three are conditional on shaderInt8, storageBuffer8BitAccess and vulkanMemoryModel being ENABLED on the logical device, exactly as #142's shaderInt64 was, and with the same failure mode: RADV computes the right answer anyway and the violation is visible only under VK_LAYER_KHRONOS_validation.

val vk_structure_type_physical_device_8bit_storage_features : int

val vk_structure_type_physical_device_vulkan_memory_model_features : int

val vk_structure_type_physical_device_subgroup_properties : int

val vk_structure_type_physical_device_driver_properties : int

val vk_structure_type_physical_device_cooperative_matrix_features_khr : int

val vk_structure_type_cooperative_matrix_properties_khr : int

val vk_khr_cooperative_matrix_extension_name : string

VK_KHR_cooperative_matrix. The extension NAME is what gates the whole path: see Vulkan_api_device.probe_coopmat for the measurement that makes checking it mandatory rather than tidy.

val vk_khr_shader_float16_int8_extension_name : string

val vk_khr_16bit_storage_extension_name : string

val vk_khr_8bit_storage_extension_name : string

val vk_khr_vulkan_memory_model_extension_name : string

type vk_extension_properties

VkExtensionProperties: char extensionName[256] then uint32 specVersion.

val vk_extension_properties : vk_extension_properties Ctypes.structure Ctypes.typ

val ext_props_extensionName : (char Ctypes_static.carray, vk_extension_properties Ctypes.structure) Ctypes.field

val ext_props_specVersion : (Unsigned.uint32, vk_extension_properties Ctypes.structure) Ctypes.field

type vk_physical_device_features_2

VkPhysicalDeviceFeatures2. features is embedded BY VALUE, and that is safe here only because vk_physical_device_features is modelled exactly — all 55 VkBool32 and nothing omitted — unlike the properties struct.

val vk_physical_device_features_2 : vk_physical_device_features_2 Ctypes.structure Ctypes.typ

val features2_sType : (Unsigned.uint32, vk_physical_device_features_2 Ctypes.structure) Ctypes.field

val features2_pNext : (unit Ctypes_static.ptr, vk_physical_device_features_2 Ctypes.structure) Ctypes.field

val features2_features : (vk_physical_device_features Ctypes.structure, vk_physical_device_features_2 Ctypes.structure) Ctypes.field

type vk_physical_device_properties_2

VkPhysicalDeviceProperties2, with the embedded VkPhysicalDeviceProperties modelled as an opaque blob.

The blob must be at least sizeof(VkPhysicalDeviceProperties), because the driver writes the whole thing. That is 4 x uint32 + deviceType + deviceName256 + pipelineCacheUUID16 + VkPhysicalDeviceLimits (504) + VkPhysicalDeviceSparseProperties (20) ~= 824 bytes; 1024 is comfortably past it and costs nothing, since exactly one of these is made per device query. An array _ char has alignment 1, and the blob's offset (16, after a 4-byte sType and an 8-aligned pNext) is already 8-aligned, so the layout matches the C struct without a named-field model of the limits.

val vk_physical_device_properties_2_blob_bytes : int

val vk_physical_device_properties_2 : vk_physical_device_properties_2 Ctypes.structure Ctypes.typ

val properties2_sType : (Unsigned.uint32, vk_physical_device_properties_2 Ctypes.structure) Ctypes.field

val properties2_pNext : (unit Ctypes_static.ptr, vk_physical_device_properties_2 Ctypes.structure) Ctypes.field

val properties2_properties : (char Ctypes_static.carray, vk_physical_device_properties_2 Ctypes.structure) Ctypes.field

type vk_physical_device_shader_float16_int8_features

VkPhysicalDeviceShaderFloat16Int8Features

val vk_physical_device_shader_float16_int8_features : vk_physical_device_shader_float16_int8_features Ctypes.structure Ctypes.typ

val f16i8_sType : (Unsigned.uint32, vk_physical_device_shader_float16_int8_features Ctypes.structure) Ctypes.field

val f16i8_pNext : (unit Ctypes_static.ptr, vk_physical_device_shader_float16_int8_features Ctypes.structure) Ctypes.field

val f16i8_shaderFloat16 : (Unsigned.uint32, vk_physical_device_shader_float16_int8_features Ctypes.structure) Ctypes.field

val f16i8_shaderInt8 : (Unsigned.uint32, vk_physical_device_shader_float16_int8_features Ctypes.structure) Ctypes.field

type vk_physical_device_16bit_storage_features

VkPhysicalDevice16BitStorageFeatures

val vk_physical_device_16bit_storage_features : vk_physical_device_16bit_storage_features Ctypes.structure Ctypes.typ

val storage16_sType : (Unsigned.uint32, vk_physical_device_16bit_storage_features Ctypes.structure) Ctypes.field

val storage16_pNext : (unit Ctypes_static.ptr, vk_physical_device_16bit_storage_features Ctypes.structure) Ctypes.field

val storage16_storageBuffer16BitAccess : (Unsigned.uint32, vk_physical_device_16bit_storage_features Ctypes.structure) Ctypes.field

val storage16_uniformAndStorageBuffer16BitAccess : (Unsigned.uint32, vk_physical_device_16bit_storage_features Ctypes.structure) Ctypes.field

val storage16_storagePushConstant16 : (Unsigned.uint32, vk_physical_device_16bit_storage_features Ctypes.structure) Ctypes.field

val storage16_storageInputOutput16 : (Unsigned.uint32, vk_physical_device_16bit_storage_features Ctypes.structure) Ctypes.field

type vk_physical_device_8bit_storage_features

VkPhysicalDevice8BitStorageFeatures — backlog-62 slice 3.

Field order and count are pinned against vulkan_core.h; the struct is passed by ADDRESS in a pNext chain, so a short or misordered layout is read past its end by the driver rather than diagnosed. Three feature booleans, the last two of which Sarek never requests.

val vk_physical_device_8bit_storage_features : vk_physical_device_8bit_storage_features Ctypes.structure Ctypes.typ

val storage8_sType : (Unsigned.uint32, vk_physical_device_8bit_storage_features Ctypes.structure) Ctypes.field

val storage8_pNext : (unit Ctypes_static.ptr, vk_physical_device_8bit_storage_features Ctypes.structure) Ctypes.field

val storage8_storageBuffer8BitAccess : (Unsigned.uint32, vk_physical_device_8bit_storage_features Ctypes.structure) Ctypes.field

val storage8_uniformAndStorageBuffer8BitAccess : (Unsigned.uint32, vk_physical_device_8bit_storage_features Ctypes.structure) Ctypes.field

val storage8_storagePushConstant8 : (Unsigned.uint32, vk_physical_device_8bit_storage_features Ctypes.structure) Ctypes.field

type vk_physical_device_vulkan_memory_model_features

VkPhysicalDeviceVulkanMemoryModelFeatures — backlog-62 slice 3.

GL_KHR_memory_scope_semantics is a hard prerequisite of GL_KHR_cooperative_matrix in glslang, and it lowers to OpCapability VulkanMemoryModel. Without vulkanMemoryModel enabled on the logical device, a cooperative-matrix shader is a specification violation however correct its results look.

val vk_physical_device_vulkan_memory_model_features : vk_physical_device_vulkan_memory_model_features Ctypes.structure Ctypes.typ

val memmodel_sType : (Unsigned.uint32, vk_physical_device_vulkan_memory_model_features Ctypes.structure) Ctypes.field

val memmodel_pNext : (unit Ctypes_static.ptr, vk_physical_device_vulkan_memory_model_features Ctypes.structure) Ctypes.field

val memmodel_vulkanMemoryModel : (Unsigned.uint32, vk_physical_device_vulkan_memory_model_features Ctypes.structure) Ctypes.field

val memmodel_vulkanMemoryModelDeviceScope : (Unsigned.uint32, vk_physical_device_vulkan_memory_model_features Ctypes.structure) Ctypes.field

val memmodel_vulkanMemoryModelAvailabilityVisibilityChains : (Unsigned.uint32, vk_physical_device_vulkan_memory_model_features Ctypes.structure) Ctypes.field

type vk_physical_device_cooperative_matrix_features

VkPhysicalDeviceCooperativeMatrixFeaturesKHR

val vk_physical_device_cooperative_matrix_features : vk_physical_device_cooperative_matrix_features Ctypes.structure Ctypes.typ

val coopmat_feat_sType : (Unsigned.uint32, vk_physical_device_cooperative_matrix_features Ctypes.structure) Ctypes.field

val coopmat_feat_pNext : (unit Ctypes_static.ptr, vk_physical_device_cooperative_matrix_features Ctypes.structure) Ctypes.field

val coopmat_feat_cooperativeMatrix : (Unsigned.uint32, vk_physical_device_cooperative_matrix_features Ctypes.structure) Ctypes.field

val coopmat_feat_robustBufferAccess : (Unsigned.uint32, vk_physical_device_cooperative_matrix_features Ctypes.structure) Ctypes.field

type vk_physical_device_subgroup_properties

VkPhysicalDeviceSubgroupProperties. subgroupSize is the one field this backend reads, and it replaces a hard-coded 32 in Vulkan_plugin_base.

val vk_physical_device_subgroup_properties : vk_physical_device_subgroup_properties Ctypes.structure Ctypes.typ

val subgroup_props_sType : (Unsigned.uint32, vk_physical_device_subgroup_properties Ctypes.structure) Ctypes.field

val subgroup_props_pNext : (unit Ctypes_static.ptr, vk_physical_device_subgroup_properties Ctypes.structure) Ctypes.field

val subgroup_props_subgroupSize : (Unsigned.uint32, vk_physical_device_subgroup_properties Ctypes.structure) Ctypes.field

val subgroup_props_supportedStages : (Unsigned.uint32, vk_physical_device_subgroup_properties Ctypes.structure) Ctypes.field

val subgroup_props_supportedOperations : (Unsigned.uint32, vk_physical_device_subgroup_properties Ctypes.structure) Ctypes.field

val subgroup_props_quadOperationsInAllStages : (Unsigned.uint32, vk_physical_device_subgroup_properties Ctypes.structure) Ctypes.field

type vk_physical_device_driver_properties

VkPhysicalDeviceDriverProperties. This is the driver KEY that docs/fp-contraction-policy.md 11.7 asks for: driverID is an enumerant (3 = VK_DRIVER_ID_MESA_RADV), and driverInfo carries the Mesa version string, so an allowlist can be keyed on a driver rather than on a substring of a device name.

val vk_physical_device_driver_properties : vk_physical_device_driver_properties Ctypes.structure Ctypes.typ

val driver_props_sType : (Unsigned.uint32, vk_physical_device_driver_properties Ctypes.structure) Ctypes.field

val driver_props_pNext : (unit Ctypes_static.ptr, vk_physical_device_driver_properties Ctypes.structure) Ctypes.field

val driver_props_driverID : (Unsigned.uint32, vk_physical_device_driver_properties Ctypes.structure) Ctypes.field

val driver_props_driverName : (char Ctypes_static.carray, vk_physical_device_driver_properties Ctypes.structure) Ctypes.field

val driver_props_driverInfo : (char Ctypes_static.carray, vk_physical_device_driver_properties Ctypes.structure) Ctypes.field

val driver_props_conformanceVersion : (Unsigned.uint8 Ctypes_static.carray, vk_physical_device_driver_properties Ctypes.structure) Ctypes.field

type vk_cooperative_matrix_properties

VkCooperativeMatrixPropertiesKHR — one advertised configuration.

AType/BType/CType/ResultType are VkComponentTypeKHR and scope is VkScopeKHR; both are C enums, hence int32_t. For the enumerant values, read the constants below — NOT a restatement here.

There was a restatement here, and it was wrong (f16 = 3, f32 = 4, s8 = 7, s32 = 9, u8 = 0, u32 = 1: the correct values rotated), in the one file whose entire narrative is a wrong enumerant table producing plausible garbage. It survived because test_vulkan_coopmat_enumerants pins the CONSTANTS and nothing pins the PROSE — the same asymmetry, one layer up. Hence no second copy of the numbers: a comment that cannot disagree with the code is the only kind worth having here.

val vk_cooperative_matrix_properties : vk_cooperative_matrix_properties Ctypes.structure Ctypes.typ

val coopmat_props_sType : (Unsigned.uint32, vk_cooperative_matrix_properties Ctypes.structure) Ctypes.field

val coopmat_props_pNext : (unit Ctypes_static.ptr, vk_cooperative_matrix_properties Ctypes.structure) Ctypes.field

val coopmat_props_MSize : (Unsigned.uint32, vk_cooperative_matrix_properties Ctypes.structure) Ctypes.field

val coopmat_props_NSize : (Unsigned.uint32, vk_cooperative_matrix_properties Ctypes.structure) Ctypes.field

val coopmat_props_KSize : (Unsigned.uint32, vk_cooperative_matrix_properties Ctypes.structure) Ctypes.field

val coopmat_props_AType : (int32, vk_cooperative_matrix_properties Ctypes.structure) Ctypes.field

val coopmat_props_BType : (int32, vk_cooperative_matrix_properties Ctypes.structure) Ctypes.field

val coopmat_props_CType : (int32, vk_cooperative_matrix_properties Ctypes.structure) Ctypes.field

val coopmat_props_ResultType : (int32, vk_cooperative_matrix_properties Ctypes.structure) Ctypes.field

val coopmat_props_saturatingAccumulation : (Unsigned.uint32, vk_cooperative_matrix_properties Ctypes.structure) Ctypes.field

val coopmat_props_scope : (int32, vk_cooperative_matrix_properties Ctypes.structure) Ctypes.field

VkComponentTypeKHR and VkScopeKHR enumerants

Verbatim from Khronos vulkan_core.h. These are NOT in a memorable order — the float types come first and the integer types are interleaved by width rather than by signedness — and an earlier draft of this file guessed them wrong in a way that produced entirely plausible garbage: f16 * s8 -> s32 and u8 * u8 -> u8 configurations that no hardware advertises, with eight of the fourteen real ones silently dropped as unrepresentable. Nothing about the resulting values looked malformed. They are pinned by test_vulkan_coopmat_enumerants against the same source.

val vk_component_type_float16 : int32

val vk_component_type_float32 : int32

val vk_component_type_float64 : int32

val vk_component_type_sint8 : int32

val vk_component_type_sint16 : int32

val vk_component_type_sint32 : int32

val vk_component_type_sint64 : int32

val vk_component_type_uint8 : int32

val vk_component_type_uint16 : int32

val vk_component_type_uint32 : int32

val vk_component_type_uint64 : int32

val vk_scope_device : int32

val vk_scope_workgroup : int32

val vk_scope_subgroup : int32

val vk_scope_queue_family : int32